Séparation UserHistoryGrid dans un component
