+++ DataTypes/UserHistory/UserHistoryGrid.module.scss
@@ -1,0 +1,16 @@
+
+
+.history-table {
+    display: grid;
+    grid-template-columns: 25% 40% 10% 10%;
+
+    //align-content: center;
+    > div{
+        background-color: var(--grey);
+        outline:2px solid black;
+        //border: 2px solid black;
+        padding: 5px;
+        font-size: 15px;
+        text-align: center;
+    }
+}
